Fixes warning in Vector2::rotate() about double.

+++ gameplay-encoder/src/Vector2.cpp
@@ -209,8 +209,8 @@ void Vector2::scale(const Vector2& scale)
 
 void Vector2::rotate(const Vector2& point, float angle)
 {
-    double sinAngle = sin(angle);
-    double cosAngle = cos(angle);
+    float sinAngle = sin(angle);
+    float cosAngle = cos(angle);
 
     if (point.isZero())
     {
